Endret pakkenavn for å reflektere appens nye navn, oppdaterte readme.
