+++ src/main/java/no/nav/brukernotifikasjon/schemas/internal/builders/StatusoppdateringBuilder.java
@@ -1,9 +1,9 @@
-package no.nav.brukernotifikasjon.schemas.builders;
+package no.nav.brukernotifikasjon.schemas.internal.builders;
 
 import no.nav.brukernotifikasjon.schemas.Statusoppdatering;
-import no.nav.brukernotifikasjon.schemas.builders.domain.Eventtype;
-import no.nav.brukernotifikasjon.schemas.builders.domain.StatusGlobal;
-import no.nav.brukernotifikasjon.schemas.builders.util.ValidationUtil;
+import no.nav.brukernotifikasjon.schemas.internal.builders.domain.Eventtype;
+import no.nav.brukernotifikasjon.schemas.internal.builders.domain.StatusGlobal;
+import no.nav.brukernotifikasjon.schemas.internal.builders.util.ValidationUtil;
 
 import java.net.URL;
 import java.time.LocalDateTime;
