+++ src/main/java/no/nav/brukernotifikasjon/schemas/internal/builders/domain/StatusGlobal.java
@@ -1,4 +1,4 @@
-package no.nav.brukernotifikasjon.schemas.builders.domain;
+package no.nav.brukernotifikasjon.schemas.internal.builders.domain;
 
 public enum StatusGlobal {
     SENDT, MOTTATT, UNDER_BEHANDLING, FERDIG;
