+++ src/test/java/no/nav/brukernotifikasjon/schemas/internal/builders/OppgaveAvroTest.java
@@ -1,4 +1,4 @@
-package no.nav.brukernotifikasjon.schemas.builders;
+package no.nav.brukernotifikasjon.schemas.internal.builders;
 
 import no.nav.brukernotifikasjon.schemas.Oppgave;
 import org.junit.jupiter.api.Test;
