+++ src/test/java/no/nav/brukernotifikasjon/schemas/internal/builders/StatusoppdateringAvroTest.java
@@ -1,8 +1,7 @@
-package no.nav.brukernotifikasjon.schemas.builders;
+package no.nav.brukernotifikasjon.schemas.internal.builders;
 
-import no.nav.brukernotifikasjon.schemas.Beskjed;
 import no.nav.brukernotifikasjon.schemas.Statusoppdatering;
-import no.nav.brukernotifikasjon.schemas.builders.domain.StatusGlobal;
+import no.nav.brukernotifikasjon.schemas.internal.builders.domain.StatusGlobal;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
