+++ src/test/java/no/nav/brukernotifikasjon/schemas/internal/builders/util/ValidationUtilTest.java
@@ -1,8 +1,8 @@
-package no.nav.brukernotifikasjon.schemas.builders.util;
+package no.nav.brukernotifikasjon.schemas.internal.builders.util;
 
-import no.nav.brukernotifikasjon.schemas.builders.domain.Eventtype;
-import no.nav.brukernotifikasjon.schemas.builders.exception.FieldValidationException;
-import no.nav.brukernotifikasjon.schemas.builders.exception.UnknownEventtypeException;
+import no.nav.brukernotifikasjon.schemas.internal.builders.domain.Eventtype;
+import no.nav.brukernotifikasjon.schemas.internal.builders.exception.FieldValidationException;
+import no.nav.brukernotifikasjon.schemas.internal.builders.exception.UnknownEventtypeException;
 import org.junit.jupiter.api.Test;
 
 import java.net.URL;
